feat(demo-integrations): refactor of `Dialogs + browser back navigation` tests
